build: update target framework to net9.0 (#190)
* build: update target framework to net9.0

Updated the target framework to net9.0 in the project files and Dockerfile for compatibility with the latest features and improvements.

* refactor: Remove line-height and add padding in CSS

- Removed line-height property and added padding to improve the styling of the footer in the CSS file.

feat: Update footer information in Razor view

- Updated the footer information in the Razor view to include additional details such as disclaimer, technology used, and theme attribution.

build: Update NuGet package references

- Updated the NuGet package references for various dependencies to the latest preview versions for compatibility and feature enhancements.

* build: Update .NET SDK and ASP.NET image versions

Update the .NET SDK version to 9.0-preview-alpine for faster debugging
and development. Also, update the ASP.NET image version to
9.0-preview-alpine for improved compatibility. These changes ensure
the project stays up to date with the latest technologies.