feat: Add MVP contributor ID to external links
This commit adds functionality to attach an MVP contributor ID to external links that match specific domains. It also includes a helper method, AttachMvpContributorId, which handles the attachment of the ID to the links.

The MVP contributor ID is set to "DT-MVP-5000879" (Gordon Beeming).

The domains that are considered for attaching the MVP contributor ID are:
- docs.microsoft.com
- learn.microsoft.com
- social.technet.microsoft.com
- azure.microsoft.com
- techcommunity.microsoft.com
- social.msdn.microsoft.com
- devblogs.microsoft.com
- developer.microsoft.com
- channel9.msdn.com
- gallery.technet.microsoft.com
- cloudblogs.microsoft.com
- technet.microsoft.com
- docs.azure.cn
- www.azure.cn
- msdn.microsoft.com
- blogs.msdn.microsoft.com
- blogs.technet.microsoft.com
- microsoft.com/handsonlabs
- csc.docs.microsoft.com
- code.visualstudio.com

The AttachMvpContributorId method is copied and altered from Jesse Houwing's code.